908: normalize strings better
