+++ src/controllers/enrollmentReports/mapUtils/enum.ts
@@ -1,4 +1,4 @@
-import { normalizeString, processedStringsMatch } from '.';
+import { normalizeString } from '../../../utils/normalizeString';
 import {
   Gender,
   AgeGroup,
@@ -60,11 +60,9 @@ export const mapEnum = <T>(
     // the acronym (i.e. CDC), full name (i.e. Child day care)
     // or full combined (i.e. CDC - Child day care) version
     if (!ret && opts.isFundingSource) {
-      let normalizedReferences = normalizeString(refString, true);
-      if (!Array.isArray(normalizedReferences)) {
-        // Make typescript happy
-        normalizedReferences = [normalizedReferences];
-      }
+      let normalizedReferences = refString
+        .split('-')
+        .map((stringPart) => normalizeString(stringPart));
       normalizedReferences.forEach((_normalizedReference) => {
         if (processedStringsMatch(normalizedInput, _normalizedReference)) {
           ret = ref;
@@ -77,3 +75,10 @@ export const mapEnum = <T>(
   }
   return ret;
 };
+
+function processedStringsMatch(str1: string, str2: string) {
+  if (str1.startsWith(str2) || str2.startsWith(str1)) {
+    return true;
+  }
+  return false;
+}
